refactor React file structure, double quotation, semi-colons added
