--- client/src/main/ArtifactItem.js
@@ -1,8 +1,0 @@
-import React from 'react'
-
-const ArtifactItem = props => {
-    return ( <div className="container"></div>)
-}
-
-
-export default ArtifactItem
